Move mutable buffer functions into a typeclass.
There are two instances of the typeclass supplied: one for Effect +
EffectBuffer and one for ST + STBuffer.

This makes it possible to write code that mutates buffers that can
run in either Effect or ST.

The functional dependencies on MutableBuffer go in both directions so
that the type-checker only needs to know _either_ the type of the buffer
or the type of the effect: this should make it easier to write generic
code without type assertions, but it does mean that both typeclass
instances must go in the same module (Node.Buffer.Mutable) in order to
avoid orphaned instances.